Refactor file structure and main.rb to be solid
