+++ classes/teacher.rb
@@ -1,4 +1,5 @@
-require './person'
+require 'pry'
+require './classes/person'
 
 class Teacher < Person
   attr_accessor :specialization, :id
