+++ decorators/trimmer_decorator.rb
@@ -1,4 +1,4 @@
-require './decorator'
+require './decorators/decorator'
 
 class TrimmerDecorator < BaseDecorator
   def correct_name
